requirements.txt
openai==1.3.5
elevenlabs==0.2.24
ffmpeg-python==0.2.0
python-dotenv==1.0.0
requests==2.31.0
